Fix RBE build for oneAPI (#130)
This PR fixes the issue 
icpx: error: unable to execute command: Executable "spirv-to-ir-wrapper"
doesn't exist!

We see the above error when we try to run the presubmit job: XLA Linux
x86 GPU oneAPI on the RBE platform.
We have submitted a PR to run the above presubmit job on the RBE
platform: https://github.com/openxla/xla/pull/33388